+++ client/src/components/Research.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react'
-import { researchData, currentResearch } from '../data/data.js'
+import { currentResearch } from '../data/data.js'
 
 import research from '../assets/research2.jpg'
-import { Btn } from './WhatIDo'
 import { Background, Title } from './Writing'
 import styled from 'styled-components'
+import ResearchArticles from './ResearchArticles.jsx'
 
 
 const BackgroundImage = styled(Background)`
@@ -35,38 +35,13 @@ const ResearchSection = styled.section`
   background-color: rgb(14,24,84);
   padding: 40px 0;
 `
-const ResearchDetail = styled.article`
-  color: white;
-  font-family: 'Manrope', sans-serif;
-  font-weight: 700;
-  line-height: 1.75;
-  letter-spacing: 1.75px;
-  font-size: 24px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 30px 0;
-`
-const Publication = styled.p`
-  width: 50%;
-  margin: 20px 0;
-`
+
 export default function Research() {
 
   useEffect(() => {
     window.scrollTo(0, 0)
   }, [])
 
-  const research = researchData.map((research, idx) => {
-    return (
-      <ResearchDetail key={idx}>
-        <Publication>{research.publication}</Publication>
-        <a href={research.url} target='_blank' rel='noopener noreferrer'>
-          <Btn>Go to</Btn>
-        </a>
-      </ResearchDetail>
-    )
-  })
 
   const currenResearchDescription = currentResearch.map((paragraph, idx) => {
     return (
@@ -84,7 +59,7 @@ export default function Research() {
         </CurrentResearch>
       </BackgroundImage>
       <ResearchSection>
-        {research}
+        <ResearchArticles />
       </ResearchSection>
     </div>
   )
